\name{EIPs-package}
\alias{EIPs-package}
\alias{EIPs}
\docType{package}
\title{
Network data and functions for Eco-Industrial Parks(EIPs)
}
\description{
Network data and functions for Eco-Industrial Parks(EIPs)
}
\details{
\tabular{ll}{
Package: \tab EIPs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-16\cr
License: \tab GPL\cr
}
Currently mainly the network data of 16 EIPs, and the functions to generate igraph or statnet network objects.
}
\author{
Yang LI <sdshiyan2005@gmail.com>
}
\references{
use data("eipdatasource") to check the name and data source of EIPs.
}
