% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBMRAlgorithm.r
\name{EIPS}
\alias{EIPS}
\title{EBMRAlgorithm Class}
\description{
The \code{EIPS} class provides an implementation for estimating
the coefficients in a propensity score model and related functionalities
based on the ensemble method for inverse probability weighting (IPW).
}
\section{Public Methods}{

The following methods are available in the EIPS class:

\code{initialize(y_names, ps_specifications, data, W, wt = NULL, bounds = 5)}:
Initializes the class and estimates the propensity scores.

\code{WangShaoKim2014(formula, h_alpha, inv_link, init = NULL)}:
Implements the Wang, Shao, and Kim (2014) approach.

\code{EBMR_IPW(h_nu)}:
Computes the IPW estimator.
}

\section{Private Fields}{

\describe{
\item{\code{r}}{The response indicator.}
\item{\code{y}}{The outcome variable(s) used in the model.}
\item{\code{n}}{The number of observations.}
\item{\code{ps_fit.list}}{List of fitted propensity score models.}
}
}

\examples{
\dontrun{
ps_specifications = list(
  formula.list = list(formula1, formula2),
  h_alpha.list = list(function(d) d[c("x1","x2")], c("x1")),
  inv_link = inv_link_function
)

ebmr = EIPS$new("y", ps_specifications, data, W)
}

}
